데이터 수집 코드 변경

+++ src/main/java/devy/americano/backend/service/NewsService.java
@@ -2,13 +2,16 @@ package devy.americano.backend.service;
 
 import devy.americano.backend.domain.News;
 import devy.americano.backend.domain.NewsContents;
+import devy.americano.backend.domain.PublisherRss;
 import devy.americano.backend.mapper.NewsMapper;
+import devy.americano.backend.mapper.PublisherRssMapper;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.List;
 
 /**
@@ -18,10 +21,39 @@ import java.util.List;
 @Service
 public class NewsService {
     private final Logger logger = LoggerFactory.getLogger(NewsService.class);
+    private final PublisherRssMapper publisherRssMapper;
     private final NewsMapper newsMapper;
+
+    private boolean inAddingNews = false;
 
     public List<News> newsList() {
         return newsMapper.selectAllNews();
+    }
+
+    public void newsCrawling() {
+        if(inAddingNews) {
+            return;
+        }
+
+        List<News> newsList = newsMapper.selectNewsNotYet();
+        if(newsList.size() == 0) {
+            return;
+        }
+
+        News news = newsList.get(0);
+        PublisherRss publisherRss = publisherRssMapper.selectPublisherRssByNo(news.getPublisherRssNo());
+
+        if(news.getLink() != null && !news.getLink().trim().isEmpty()) {
+            try {
+                logger.info(news.getLink());
+                NewsCrawler.newsCrawler(publisherRss, news).image().author().pubDate();
+                news.setRegDateLDT(LocalDateTime.now());
+                newsMapper.updateNews(news);
+            } catch(Exception e) {
+                logger.info("Error to Crawling : " + news.getLink());
+                e.printStackTrace();
+            }
+        }
     }
 
     public List<NewsContents> searchNewsContents(int pageNo) {
@@ -35,14 +67,23 @@ public class NewsService {
             return;
         }
 
+        inAddingNews = true;
+
         for(News news : newsList) {
             newsMapper.insertNews(news);
         }
+
+        inAddingNews = false;
     }
 
     public void removeOldNews() {
         int removeBefore = 6;
         LocalDateTime date = LocalDateTime.now().minusHours(9 + removeBefore);
+
+        if(inAddingNews) {
+            return;
+        }
+
         newsMapper.deleteOldNews(date);
     }
